Drop windowHelpers and gate the veto unregistration like its registration
The shell window helper moves into namedPipe, its only user, so the module
that held the deleted full screen check goes away.
